+++ changelog.md
@@ -5,6 +5,7 @@
 - Changed several recipes to use oredict ingredients
 - Moved Void Walker Boots taint immunity to event (should fix some jank)
 - Improved keybinding handling (should fix edge-case flying carpet conflicts)
+- Improved Wapr ring potion removal (should fix some edge-case damage events)
 
 ## v2.2.3
 ------